move generateMerkleRootAndProof into rootchain_helpers
